remove sympy to function code from expression.py to sym_to_func.py
