+++ ppsci/utils/expression.py
@@ -12,24 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import functools
 from typing import TYPE_CHECKING
 from typing import Callable
 from typing import Dict
-from typing import List
 from typing import Optional
 from typing import Tuple
-from typing import Union
 
 import paddle
-import sympy as sp
 from paddle import jit
 from paddle import nn
-from typing_extensions import TypeAlias
-
-from ppsci.autodiff import hessian
-from ppsci.autodiff import jacobian
-from ppsci.utils import logger
 
 if TYPE_CHECKING:
     from ppsci import constraint
@@ -200,246 +191,3 @@ class ExpressionSolver(nn.Layer):
             clear()
 
         return output_dict
-
-
-FUNC_MAP = {
-    sp.sin: paddle.sin,
-    sp.cos: paddle.cos,
-    sp.exp: paddle.exp,
-    sp.Pow: paddle.pow,
-    sp.log: paddle.log,
-    sp.tan: paddle.tan,
-    sp.Max: paddle.maximum,
-    sp.Min: paddle.minimum,
-    sp.Abs: paddle.abs,
-    sp.Heaviside: functools.partial(paddle.heaviside, y=paddle.zeros([])),
-}
-
-SYMPY_BUILTIN_FUNC: TypeAlias = Union[
-    sp.sin,
-    sp.cos,
-    sp.exp,
-    sp.Pow,
-    sp.log,
-    sp.tan,
-    sp.Max,
-    sp.Min,
-    sp.Abs,
-    sp.Heaviside,
-]
-
-
-def _single_derivate_func(dvar: paddle.Tensor, invar: paddle.Tensor, order: int):
-    order_left = order
-    while order_left > 0:
-        if order_left >= 2:
-            dvar = hessian(dvar, invar)
-            order_left -= 2
-        else:
-            dvar = jacobian(dvar, invar)
-            order_left -= 1
-    return dvar
-
-
-def _cvt_to_key(sympy_node: sp.Basic):
-    if isinstance(
-        sympy_node, (sp.Symbol, sp.core.function.UndefinedFunction, sp.Function)
-    ):
-        if hasattr(sympy_node, "name"):
-            # custom function
-            return sympy_node.name
-        else:
-            str(sympy_node)
-    elif isinstance(sympy_node, sp.Derivative):
-        # convert Derivative(u(x,y),(x,2),(y,2)) to "u__x__x__y__y"
-        expr_str = sympy_node.args[0].name
-        for symbol, order in sympy_node.args[1:]:
-            expr_str += f"__{symbol}" * order
-        return expr_str
-    else:
-        return str(sympy_node)
-
-
-class Node(nn.Layer):
-    """The base class of the node in expression tree."""
-
-    def __init__(self, expr: sp.Basic):
-        super().__init__()
-        self.expr = expr
-        self.key = _cvt_to_key(self.expr)
-
-    def forward(self, **kwargs):
-        raise NotImplementedError("Node.forward is not implemented")
-
-    def __str__(self):
-        return (
-            self.__class__.__name__ + f"(expr: {self.expr}), type: {type(self.expr)})"
-        )
-
-
-class OperatorNode(Node):
-    """
-    A node representing a sp operator in the computational graph.
-    """
-
-    def __init__(self, expr: SYMPY_BUILTIN_FUNC):
-        super().__init__(expr)
-
-    def forward(self, data_dict: Dict):
-        if self.expr.func == sp.Add:
-            data_dict[self.key] = sum(
-                [data_dict[_cvt_to_key(arg)] for arg in self.expr.args]
-            )
-        elif self.expr.func == sp.Mul:
-            data_dict[self.key] = data_dict[_cvt_to_key(self.expr.args[0])]
-            for arg in self.expr.args[1:]:
-                data_dict[self.key] = data_dict[self.key] * data_dict[_cvt_to_key(arg)]
-        elif self.expr.func == sp.Derivative:
-            if self.key in data_dict:
-                return data_dict
-            data_dict[self.key] = data_dict[_cvt_to_key(self.expr.args[0])]
-            for symbol, order in self.expr.args[1:]:
-                data_dict[self.key] = _single_derivate_func(
-                    data_dict[self.key],
-                    data_dict[_cvt_to_key(symbol)],
-                    order,
-                )
-        else:
-            try:
-                func = FUNC_MAP[self.expr.func]
-            except KeyError:
-                raise NotImplementedError(
-                    f"'{self.expr.func}' operator is not supported now."
-                )
-            if self.expr.func == sp.Heaviside:
-                data_dict[self.key] = func(data_dict[_cvt_to_key(self.expr.args[0])])
-            else:
-                data_dict[self.key] = func(
-                    *[data_dict[_cvt_to_key(arg)] for arg in self.expr.args]
-                )
-        return data_dict
-
-
-class LayerNode(Node):
-    """
-    A node representing a neural network in the computational graph
-    """
-
-    def __init__(self, expr: sp.core.function.UndefinedFunction, model: nn.Layer):
-        super().__init__(expr)
-        self.model = model
-
-    def forward(self, data_dict: Dict):
-        if self.key in data_dict:
-            return data_dict
-
-        output_dict = self.model(data_dict)
-        print("call model forward")
-        for key, value in output_dict.items():
-            data_dict[key] = value
-
-        return data_dict
-
-
-class ConstantNode(Node):
-    """
-    A node representing a constant in the computational graph.
-    """
-
-    def __init__(self, expr: Union[sp.Number, sp.NumberSymbol]):
-        super().__init__(expr)
-        if (
-            self.expr.is_Float
-            or self.expr.is_Integer
-            or self.expr.is_Boolean
-            or self.expr.is_Rational
-        ):
-            self.expr = float(self.expr)
-        else:
-            raise TypeError(
-                f"expr({expr}) should be float/int/bool, but got {type(self.expr)}"
-            )
-        self.expr = paddle.to_tensor(self.expr)
-
-    def forward(self, data_dict: Dict):
-        data_dict[self.key] = self.expr
-        return data_dict
-
-
-class ComposedFunc(nn.Layer):
-    """
-    Compose multiple functions into one function.
-    """
-
-    def __init__(self, target: str, funcs: List[Node]):
-        super().__init__()
-        self.funcs = funcs
-        self.target = target
-
-    def forward(self, data_dict: Dict):
-        for func in self.funcs:
-            data_dict = func(data_dict)
-        return data_dict[self.funcs[-1].key]  # return the computed result of root node
-
-
-def _post_traverse(cur_node: sp.Basic, nodes: List[sp.Basic]) -> List[sp.Basic]:
-    # traverse into sub-nodes
-    if isinstance(cur_node, sp.core.function.UndefinedFunction):
-        nodes.append(cur_node)
-    elif isinstance(cur_node, sp.Function):
-        for arg in cur_node.args:
-            nodes = _post_traverse(arg, nodes)
-        nodes.append(cur_node)
-    elif isinstance(cur_node, sp.Derivative):
-        nodes = _post_traverse(cur_node.args[0], nodes)
-        nodes.append(cur_node)
-    elif isinstance(cur_node, sp.Symbol):
-        return nodes
-    elif isinstance(cur_node, sp.Number):
-        nodes.append(cur_node)
-    else:
-        for arg in cur_node.args:
-            nodes = _post_traverse(arg, nodes)
-        nodes.append(cur_node)
-    return nodes
-
-
-def sympy_to_function(target: str, expr: sp.Expr, models: nn.Layer) -> ComposedFunc:
-    """
-    Convert a sp expression to a ComposedFunc.
-    """
-    sympy_nodes = []
-    sympy_nodes = _post_traverse(expr, sympy_nodes)
-    sympy_nodes = [node for node in sympy_nodes if not node.is_Symbol]
-    sympy_nodes = list(
-        dict.fromkeys(sympy_nodes)
-    )  # remove duplicates with topo-order kept
-
-    callable_nodes = []
-    for i, node in enumerate(sympy_nodes):
-        logger.debug(f"tree node [{i + 1}/{len(sympy_nodes)}]: {node}")
-        if isinstance(node.func, sp.core.function.UndefinedFunction):
-            match = False
-            for model in models:
-                if str(node.func.name) in model.output_keys:
-                    callable_nodes.append(LayerNode(node, model))
-                    if match:
-                        raise ValueError(
-                            f"function {node} can match at least 2 output key of models, which is forbidden."
-                        )
-                    match = True
-        elif (
-            isinstance(node, tuple(FUNC_MAP.keys()))
-            or node.is_Add
-            or node.is_Mul
-            or node.is_Derivative
-            or node.is_Pow
-        ):
-            callable_nodes.append(OperatorNode(node))
-        elif node.is_Number or node.is_NumberSymbol:
-            callable_nodes.append(ConstantNode(node))
-        else:
-            raise NotImplementedError(
-                f"The node {node} is not supported in sympy_to_function."
-            )
-    return ComposedFunc(target, callable_nodes)
